--- jest.config..js.txt
@@ -1,6 +1,0 @@
-module.exports = {
-  // Other Jest configuration options...
-  runner: 'jest-runner-eslint',
-  displayName: 'lint',
-  testMatch: ['<rootDir>/test/*.js'],
-};
